+++ packages/json-rpc/logs-provider/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./logs-provider"
 export * from "./with-logs-recorder"
+export * from "./public-types"